Added the ITG2 Eval screen grades, remade the animations and made the quint "model".
Added all the assets needed for result screen 3d models including making the quint.
Made SL load the models instead of the flat textures. Remade the grade animations so that it looks a bit more like ITG2.
Note that the easter egg Q-grade is not supported, and will likely never be.
Moved the logic for loading grades to it's own file to be able to do it later in the sequence, this makes it so that it will load over the statistics as this was a problem.